requirements.txt
playwright~=1.50.0
pandas~=2.2.3
python-dotenv==1.0.0
loguru~=0.7.3